abstract value formatting into its own service
